chore(facets): Fix broken styling/collapse toggles

+++ app/scripts/components/facets/facets.controllers.ts
@@ -4,19 +4,32 @@ module ngApp.components.facets.directives {
   import IFacetService = ngApp.components.facets.services.IFacetService;
 
   interface ITermsController {
-    add(facet: string, term: string): void;
+    add(facet: string, term: string, event: any): void;
+    remove(facet: string, term: string, event: any): void;
     actives: string[];
     inactives: string[];
+    displayCount: number;
+    originalDisplayCount: number;
+    collapsed: boolean;
+    expanded: boolean;
+    toggle(event: any, property: string): void;
   }
 
   class TermsController implements ITermsController {
     title: string = "";
     name: string = "";
+    displayCount: number = 5;
+    originalDisplayCount: number = 5;
+    collapsed: boolean = false;
+    expanded: boolean = false;
     actives: string[] = [];
     inactives: string[] = [];
 
     /* @ngInject */
     constructor($scope: IFacetScope, private FacetService: IFacetService) {
+      this.collapsed = !!$scope.collapsed;
+      this.expanded = !!$scope.expanded;
+      this.displayCount = this.originalDisplayCount = $scope.displayCount || 5;
       this.title = $scope.title;
       if ($scope.facet) {
         this.refresh($scope.facet.buckets);
@@ -32,17 +45,35 @@ module ngApp.components.facets.directives {
       });
     }
 
-    add(facet: string, term: string): void {
-      this.FacetService.addTerm(facet, term);
+    add(facet: string, term: string, event: any): void {
+      if (event.which === 1 || event.which === 13) {
+        this.FacetService.addTerm(facet, term);
+      }
     }
 
-    remove(facet: string, term: string): void {
-      this.FacetService.removeTerm(facet, term);
+    remove(facet: string, term: string, event: any): void {
+      if (event.which === 1 || event.which === 13) {
+        this.FacetService.removeTerm(facet, term);
+      }
     }
 
     refresh(terms) {
       this.actives = this.FacetService.getActives(this.name, terms);
       this.inactives = _.difference(terms, this.actives);
+    }
+
+    toggle(event: any, property: string) {
+      if (event.which === 1 || event.which === 13) {
+        this[property] = !this[property];
+      }
+
+      if (property === "collapsed") {
+        angular.element(event.target).attr("aria-collapsed", this.collapsed.toString());
+      }
+
+      if (property === "expanded") {
+        this.displayCount = this.expanded ? this.inactives.length : this.originalDisplayCount;
+      }
     }
   }
